Move TaskList to storage package
