+++ src/main/java/tick/commands/Command.java
@@ -2,7 +2,7 @@ package tick.commands;
 
 import tick.exceptions.TickException;
 import tick.storage.Storage;
-import tick.tasks.TaskList;
+import tick.storage.TaskList;
 import tick.ui.Ui;
 
 public abstract class Command {
